requirements.txt
flask
opencv-python
pillow
torch
torchvision
numpy
pandas
scikit-learn
plotly
